requirements.txt
Flask==0.10.1
docker-py
pyyaml
